DROTH-3922 Moved a method from filler to asset class
